requirements.txt
requests~=2.32.3
azure-core~=1.33.0